Deleted BiomassRemoval interface.  Needed bits put in organ interface
