+++ Models/Plant/PlantTypes.cs
@@ -160,4 +160,22 @@
         /// <summary>The crop_type</summary>
         public String crop_type = "";
     }
+    
+    ///<summary>Data passed to each organ when a biomass remove event occurs</summary>
+    [Serializable]
+    public class OrganBiomassRemovalType
+    {
+        /// <summary>
+        /// The amount of biomass to removeed from each organ on harvest, cut or graze.
+        /// </summary>
+        public double FractionRemoved { get; set; }
+        /// <summary>
+        /// The amount of biomass to removed from each organ and passed to residue pools on defoliation events
+        /// </summary>
+        public double FractionToResidue { get; set; }
+        /// <summary>
+        /// The amount of biomass to removed from each organ and passed to residue pools on defoliation events
+        /// </summary>
+        public string RemovalMethod { get; set; }
+    }
 }
